removed unused abstraction for LoginService
