+++ bot/services/__init__.py
@@ -1,3 +1,2 @@
 from .LoginService import LoginService
-from .LKLoginService import LKLoginService
 from .RoleSorting import sortRoles
